move cowswap files to types and helpers where needed; make big errors on swap&bridge silent
