+++ src/services/cowswap/api.ts
@@ -1,16 +1,26 @@
-import {
-  formatUnits,
-  getAddress,
-  Interface,
-  isAddress,
-  keccak256,
-  solidityPacked,
-  toUtf8Bytes,
-  TypedDataEncoder,
-  ZeroAddress
-} from 'ethers'
+import { getAddress, isAddress, keccak256, toUtf8Bytes, ZeroAddress } from 'ethers'
 
-import { FEE_COLLECTOR } from '@/consts/addresses'
+import {
+  buildAppData,
+  computeOrderUid,
+  ethFlowInterface,
+  getApiNetwork,
+  getOutputValueInUsd,
+  getProtocolFeeAmount,
+  getWrappedNativeTokenAddress,
+  isCowSwapTokenListEntry,
+  normalizeBuyTokenAddress,
+  normalizeCowSwapToken,
+  settlementInterface
+} from '@/services/cowswap/helper'
+import {
+  CenaPlatformResponse,
+  CenaTokenResponse,
+  CowSwapErrorResponse,
+  CowSwapOrderResponse,
+  CowSwapTokenListEntry,
+  CowSwapTrade
+} from '@/services/cowswap/types'
 
 import SwapAndBridgeProviderApiError from '../../classes/SwapAndBridgeProviderApiError'
 import { getTokenUsdAmount } from '../../controllers/signAccountOp/helper'
@@ -38,10 +48,8 @@ import {
   isNoFeeToken
 } from '../../libs/swapAndBridge/swapAndBridge'
 import {
+  CENA_API_BASE_URL,
   COWSWAP_API_BASE_URL,
-  COWSWAP_APP_CODE,
-  COWSWAP_APP_DATA_VERSION,
-  COWSWAP_BUY_NATIVE_TOKEN_ADDRESS,
   COWSWAP_ETH_FLOW_ADDRESS,
   COWSWAP_ORDER_VALIDITY_SECONDS,
   COWSWAP_SETTLEMENT_ADDRESS,
@@ -49,224 +57,6 @@ import {
   COWSWAP_TOKEN_LIST_URL,
   COWSWAP_VAULT_RELAYER_ADDRESS
 } from './constants'
-
-const settlementInterface = new Interface(['function setPreSignature(bytes orderUid, bool signed)'])
-const ethFlowInterface = new Interface([
-  'function createOrder((address buyToken,address receiver,uint256 sellAmount,uint256 buyAmount,bytes32 appData,uint256 feeAmount,uint32 validTo,bool partiallyFillable,int64 quoteId) order) payable returns (bytes32 orderHash)'
-])
-
-const MAX_VALID_TO = 2 ** 32 - 1
-const CENA_API_BASE_URL = 'https://cena.ambire.com'
-
-const orderTypes = {
-  Order: [
-    { name: 'sellToken', type: 'address' },
-    { name: 'buyToken', type: 'address' },
-    { name: 'receiver', type: 'address' },
-    { name: 'sellAmount', type: 'uint256' },
-    { name: 'buyAmount', type: 'uint256' },
-    { name: 'validTo', type: 'uint32' },
-    { name: 'appData', type: 'bytes32' },
-    { name: 'feeAmount', type: 'uint256' },
-    { name: 'kind', type: 'string' },
-    { name: 'partiallyFillable', type: 'bool' },
-    { name: 'sellTokenBalance', type: 'string' },
-    { name: 'buyTokenBalance', type: 'string' }
-  ]
-}
-
-type CowSwapOrderStatus = 'presignaturePending' | 'open' | 'fulfilled' | 'cancelled' | 'expired'
-
-type CowSwapOrderResponse = {
-  status: CowSwapOrderStatus
-}
-
-type CowSwapTrade = {
-  txHash?: string | null
-}
-
-type CowSwapErrorResponse = {
-  errorType?: string
-  description?: string
-  message?: string
-}
-
-type CowSwapTokenListEntry = {
-  address: string
-  chainId: number
-  decimals: number
-  logoURI?: string
-  name: string
-  symbol: string
-}
-
-type CenaPlatformResponse = {
-  platformId?: string
-}
-
-type CenaTokenResponse = {
-  blacklist?: boolean
-  decimals?: Record<string, number>
-  image?: {
-    large?: string
-    small?: string
-    thumb?: string
-  }
-  name?: string
-  platforms?: Record<string, string>
-  removed?: boolean
-  symbol?: string
-}
-
-const isCowSwapTokenListEntry = (value: unknown): value is CowSwapTokenListEntry => {
-  if (!value || typeof value !== 'object') return false
-  if (
-    !('address' in value) ||
-    !('chainId' in value) ||
-    !('decimals' in value) ||
-    !('name' in value) ||
-    !('symbol' in value)
-  )
-    return false
-
-  return (
-    typeof value.address === 'string' &&
-    isAddress(value.address) &&
-    typeof value.chainId === 'number' &&
-    Number.isInteger(value.chainId) &&
-    typeof value.decimals === 'number' &&
-    Number.isInteger(value.decimals) &&
-    value.decimals >= 0 &&
-    value.decimals <= 255 &&
-    typeof value.name === 'string' &&
-    !!value.name &&
-    typeof value.symbol === 'string' &&
-    !!value.symbol &&
-    (!('logoURI' in value) || value.logoURI === undefined || typeof value.logoURI === 'string')
-  )
-}
-
-const normalizeCowSwapToken = (token: CowSwapTokenListEntry): SwapAndBridgeToToken => ({
-  address: getAddress(token.address),
-  chainId: token.chainId,
-  decimals: token.decimals,
-  icon: token.logoURI || '',
-  name: token.name,
-  symbol: token.symbol
-})
-
-const getApiNetwork = (chainId: number) =>
-  COWSWAP_SUPPORTED_CHAINS.find((chain) => chain.chainId === chainId)?.apiNetwork
-
-const getWrappedNativeTokenAddress = (chainId: number) =>
-  COWSWAP_SUPPORTED_CHAINS.find((chain) => chain.chainId === chainId)?.wrappedNativeTokenAddress
-
-const normalizeBuyTokenAddress = (address: string) =>
-  address.toLowerCase() === ZeroAddress.toLowerCase()
-    ? COWSWAP_BUY_NATIVE_TOKEN_ADDRESS
-    : getAddress(address)
-
-const getProtocolFeeAmount = (buyAmount: bigint, protocolFeeBps: number) => {
-  if (protocolFeeBps <= 0) return 0n
-
-  const precision = 100000n
-  const protocolFeeBpsWithPrecision = BigInt(Math.round(protocolFeeBps * Number(precision)))
-  const denominator = 10000n * precision - protocolFeeBpsWithPrecision
-
-  if (denominator <= 0n) {
-    throw new SwapAndBridgeProviderApiError(
-      'Unable to fetch the quote. CoW Swap returned an invalid fee.'
-    )
-  }
-
-  return (buyAmount * protocolFeeBpsWithPrecision) / denominator
-}
-
-const buildAppData = ({ slippageBps, feeBps }: { slippageBps: number; feeBps?: number }) => {
-  const appData = {
-    appCode: COWSWAP_APP_CODE,
-    metadata: {
-      orderClass: { orderClass: 'market' },
-      ...(feeBps
-        ? {
-            partnerFee: {
-              recipient: FEE_COLLECTOR,
-              volumeBps: feeBps
-            }
-          }
-        : {}),
-      quote: { slippageBips: slippageBps }
-    },
-    version: COWSWAP_APP_DATA_VERSION
-  }
-  const fullAppData = JSON.stringify(appData)
-
-  return {
-    fullAppData,
-    appDataHash: keccak256(toUtf8Bytes(fullAppData))
-  }
-}
-
-const computeOrderUid = ({
-  chainId,
-  order,
-  owner,
-  isEthFlow
-}: {
-  chainId: number
-  order: CowSwapOrderCreation
-  owner: string
-  isEthFlow: boolean
-}) => {
-  const validTo = isEthFlow ? MAX_VALID_TO : order.validTo
-  const orderDigest = TypedDataEncoder.hash(
-    {
-      name: 'Gnosis Protocol',
-      version: 'v2',
-      chainId,
-      verifyingContract: COWSWAP_SETTLEMENT_ADDRESS
-    },
-    orderTypes,
-    {
-      sellToken: order.sellToken,
-      buyToken: order.buyToken,
-      receiver: order.receiver,
-      sellAmount: order.sellAmount,
-      buyAmount: order.buyAmount,
-      validTo,
-      appData: order.appDataHash,
-      feeAmount: order.feeAmount,
-      kind: order.kind,
-      partiallyFillable: order.partiallyFillable,
-      sellTokenBalance: order.sellTokenBalance,
-      buyTokenBalance: order.buyTokenBalance
-    }
-  )
-
-  return solidityPacked(
-    ['bytes32', 'address', 'uint32'],
-    [orderDigest, isEthFlow ? COWSWAP_ETH_FLOW_ADDRESS : owner, validTo]
-  )
-}
-
-const getOutputValueInUsd = ({
-  inputValueInUsd,
-  toAsset,
-  toAmount,
-  buyAmountBeforeFees
-}: {
-  inputValueInUsd: number
-  toAsset: SwapAndBridgeToToken
-  toAmount: string
-  buyAmountBeforeFees: bigint
-}) => {
-  const priceUSD = Number(toAsset.priceUSD || 0)
-  if (!priceUSD) {
-    return inputValueInUsd * (Number(toAmount) / Number(buyAmountBeforeFees))
-  }
-
-  return Number(formatUnits(toAmount, toAsset.decimals)) * priceUSD
-}
 
 export class CowSwapAPI implements SwapProvider {
   id = 'cowswap'
